Wrap platform errors
- Expose last errors per platform
- Replace errno with PP_IO_LAST_ERROR()
- Replace in logs
- Split socket differences into files
