+++ Sources/PartoutCore_C/include/portable/tun.h
@@ -64,18 +64,20 @@ int pp_tun_get_fd(const pp_tun tun);
 /* Return the device name or NULL if none. */
 const char *_Nullable pp_tun_name(const pp_tun tun);
 
-/* Reusable cross-platform. */
+#if !PARTOUT_WINDOWS
+/* Reusable with POSIX-like tun device. */
 static inline int pp_tun_handle_result(int ret) {
     if (ret < 0) {
-        if (PP_IO_WOULDBLOCK()) {
+        if (pp_io_wouldblock()) {
             return PPIOErrorWouldBlock;
         }
-        if (PP_IO_NOBUFS()) {
+        if (pp_io_nobufs()) {
             return PPIOErrorNoBufs;
         }
     }
     return ret;
 }
+#endif
 
 /* Tunnel controller. */
 typedef struct {
